move test-runner files
